Levels/Level 3/01-flask-crud-to-do/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Werkzeug==3.0.1
WTForms==3.1.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
mypy==1.7.1
black==23.12.1
ruff==0.1.9
python-dotenv==1.0.0
